President Trump  no way
I heard that the peace treaty of Russia was Mrs Clinton's plan. China VS Russia would be a world war? However, the West and Europe always drives Russia away. if so America with indifferent will not be able to action to anything and then U.S.A with G8 will be fallen then.
so america is being attack to china.
It is America that helped Chiang Kai-shek and made China a communist zone. Turning back to that American history, U.S.A will find out that Russia will be the G20. any way The peace treaty with Russia in Japan itself wrong.
Now in China, G5 vs. President Trump is Germany. I maybe remember that the latest news is that Germany entered the G5